resolved todos and reslove test case from 1 to 4

+++ database/migrations/2022_05_16_143445_create_orders_table.php
@@ -19,10 +19,10 @@ return new class extends Migration
             $table->foreignId('merchant_id')->constrained();
             $table->foreignId('affiliate_id')->nullable()->constrained();
             // TODO: Replace floats with the correct data types (very similar to affiliates table)
-            // same as affiliates table, we will use the same thing here, double handle such kind of values.
-            // double value always used when we are dealing in such types of calculation where every decimal points matter.
-            $table->double('subtotal');
-            $table->double('commission_owed')->default(0.00);
+            // same as affiliates table, we will use the same thing here, decimal handle such kind of values.
+            // decimal values always used when we are dealing in such types of calculation where every decimal points matter.
+            $table->decimal('subtotal', 10, 4);
+            $table->decimal('commission_owed', 10, 4)->default(0.00);
             $table->string('payout_status')->default(Order::STATUS_UNPAID);
             $table->string('discount_code')->nullable();
             $table->timestamps();
